requirements.txt
azure-identity==1.10.0
msgraph-core==0.2.2
